requirements.txt
streamlit
sentence-transformers
openai
faiss-cpu==1.7.4
numpy<2
python-dotenv
